pyproject.toml
[project]
name = "homerentals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "bottle>=0.13.2",
  "dotenv>=0.9.9",
  "mysql-connector-python>=9.2.0",
]
